requirements.txt
Flask>=2.0
google-api-python-client>=2.0
google-auth-oauthlib>=0.5
google-auth-httplib2>=0.1
google-generativeai>=0.3
python-dotenv>=0.20
pytz>=2023.3
cachetools>=5.0
